backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.1
gTTS==2.5.3
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
vosk==0.3.45
websockets==12.0
